Replace all long's as int's, except a selected few
Done with:
    sed -i -e 's/long/int/g' *.cpp *.h
    sed -i -e 's/int int/long long/g' *.cpp *.h
    sed -i -e 's/long long32/int int32/g' *.cpp *.h

Fixes #19

+++ src/CvtManager.h
@@ -42,27 +42,27 @@ class ControlValueTable {
 public:
 	ControlValueTable(void) {}
 	virtual ~ControlValueTable(void) {}
-	virtual bool Compile(TextBuffer *source, TextBuffer *prepText, bool legacyCompile, long *errPos, long *errLen, wchar_t errMsg[]) = 0;
+	virtual bool Compile(TextBuffer *source, TextBuffer *prepText, bool legacyCompile, int *errPos, int *errLen, wchar_t errMsg[]) = 0;
 	virtual bool IsControlProgramFormat(void) = 0;
 	virtual bool LinearAdvanceWidths(void) = 0;
-	virtual long LowestCvtNum(void) = 0;
-	virtual long HighestCvtNum(void) = 0;
-	virtual long LowestCvtIdx(void) = 0;
-	virtual long HighestCvtIdx(void) = 0;
-	virtual long CvtNumOf(long idx) = 0;
-	virtual long CvtIdxOf(long num) = 0; 
-	virtual bool CvtNumExists(long cvtNum) = 0;
-	virtual bool GetCvtValue(long cvtNum, short *cvtValue) = 0;
-	virtual bool CvtAttributesExist(long cvtNum) = 0;  // entered a cvt "comment"?
-	virtual bool GetCvtAttributes(long cvtNum, CharGroup *charGroup, LinkColor *linkColor, LinkDirection *linkDirection, CvtCategory *cvtCategory, bool *relative) = 0;
-	virtual bool GetAttributeStrings(long cvtNum, wchar_t charGroup[], wchar_t linkColor[], wchar_t linkDirection[], wchar_t cvtCategory[], wchar_t relative[]) = 0;
-	virtual long NumCharGroups(void) = 0;
+	virtual int LowestCvtNum(void) = 0;
+	virtual int HighestCvtNum(void) = 0;
+	virtual int LowestCvtIdx(void) = 0;
+	virtual int HighestCvtIdx(void) = 0;
+	virtual int CvtNumOf(int idx) = 0;
+	virtual int CvtIdxOf(int num) = 0; 
+	virtual bool CvtNumExists(int cvtNum) = 0;
+	virtual bool GetCvtValue(int cvtNum, short *cvtValue) = 0;
+	virtual bool CvtAttributesExist(int cvtNum) = 0;  // entered a cvt "comment"?
+	virtual bool GetCvtAttributes(int cvtNum, CharGroup *charGroup, LinkColor *linkColor, LinkDirection *linkDirection, CvtCategory *cvtCategory, bool *relative) = 0;
+	virtual bool GetAttributeStrings(int cvtNum, wchar_t charGroup[], wchar_t linkColor[], wchar_t linkDirection[], wchar_t cvtCategory[], wchar_t relative[]) = 0;
+	virtual int NumCharGroups(void) = 0;
 	virtual bool GetCharGroupString(CharGroup charGroup, wchar_t string[]) = 0;
 	virtual bool GetSpacingText(CharGroup charGroup, wchar_t spacingText[]) = 0;
-	virtual long GetBestCvtMatch(CharGroup charGroup, LinkColor linkColor, LinkDirection linkDirection, CvtCategory cvtCategory, long distance) = 0; // returns illegalCvtNum if no match
-	virtual void PutCvtBinary(long size, unsigned char data[]) = 0;
-	virtual void GetCvtBinary(long *size, unsigned char data[]) = 0;
-	virtual long GetCvtBinarySize(void) = 0;
+	virtual int GetBestCvtMatch(CharGroup charGroup, LinkColor linkColor, LinkDirection linkDirection, CvtCategory cvtCategory, int distance) = 0; // returns illegalCvtNum if no match
+	virtual void PutCvtBinary(int size, unsigned char data[]) = 0;
+	virtual void GetCvtBinary(int *size, unsigned char data[]) = 0;
+	virtual int GetCvtBinarySize(void) = 0;
 	virtual bool DumpControlValueTable(TextBuffer *text) = 0;
 	virtual bool CompileCharGroup(File *from, short platformID, unsigned char toCharGroupOfCharCode[], wchar_t errMsg[]) = 0;
 private:
